+++ packages/unixfs/src/commands/chmod.ts
@@ -1,7 +1,7 @@
 import * as dagPB from '@ipld/dag-pb'
 import { logger } from '@libp2p/logger'
 import { UnixFS } from 'ipfs-unixfs'
-import { recursive } from 'ipfs-unixfs-exporter'
+import { exporter, recursive } from 'ipfs-unixfs-exporter'
 import { importer } from 'ipfs-unixfs-importer'
 import last from 'it-last'
 import { pipe } from 'it-pipe'
@@ -31,13 +31,14 @@ export async function chmod (cid: CID, mode: number, blockstore: PutStore & GetS
         for await (const entry of recursive(resolved.cid, blockstore, options)) {
           let metadata: UnixFS
           let links: PBLink[] = []
+          const file = await exporter(entry.cid, blockstore, options)
 
-          if (entry.type === 'raw') {
+          if (file.type === 'raw') {
             // convert to UnixFS
-            metadata = new UnixFS({ type: 'file', data: entry.node })
-          } else if (entry.type === 'file' || entry.type === 'directory') {
-            metadata = entry.unixfs
-            links = entry.node.Links
+            metadata = new UnixFS({ type: 'file', data: file.node })
+          } else if (file.type === 'file' || file.type === 'directory') {
+            metadata = file.unixfs
+            links = file.node.Links
           } else {
             throw new NotUnixFSError()
           }