+++ packages/http/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [3.0.16](https://github.com/ipfs/helia/compare/http-v3.0.15...http-v3.0.16) (2026-01-16)
+
+
+### Documentation
+
+* update readme ([14e7873](https://github.com/ipfs/helia/commit/14e78732fdca2053ddc1cd2a3905b4faa4eaf916))
+
+
+### Dependencies
+
+* update delegated routing client to v6 ([#936](https://github.com/ipfs/helia/issues/936)) ([ec26650](https://github.com/ipfs/helia/commit/ec2665082163f3c87ff0f9d3b491295f5cdf49b0))
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/routers bumped from ^4.0.6 to ^5.0.0
+
+## [3.0.15](https://github.com/ipfs/helia/compare/http-v3.0.14...http-v3.0.15) (2026-01-15)
+
+
+### Bug Fixes
+
+* use non-deprecated factory function to create delegated client ([#934](https://github.com/ipfs/helia/issues/934)) ([20ba9cf](https://github.com/ipfs/helia/commit/20ba9cf6256961d2b664af0e8f48b5e9d009d834))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/block-brokers bumped from ^5.0.13 to ^5.0.14
+    * @helia/routers bumped from ^4.0.5 to ^4.0.6
+    * @helia/utils bumped from ^2.3.3 to ^2.3.4
+
 ## [3.0.14](https://github.com/ipfs/helia/compare/http-v3.0.13...http-v3.0.14) (2025-12-15)